requirements.txt
pydub
pyenchant
sly
nltk
termcolor
pandas
numpy>=1.12.0
pyyaml>=4.2b1
scipy>=0.18.1
TextGrid>=1.4